Feature/attributes page (#53)
* TRIB-33-minor-fixes: changes name of List<PhraseDTO> to list (#48)

* TRIB-33-minor-fixes: changes name of List<PhraseDTO> to list

* TRIB-33-minor-fixes: changes list name to phrases

* TRIB-72: Changes get attributes endpoint to return optional of array list only (#49)

* mrsbluerose/attribute-service-retool: removes private final variables and constructor. Autowires repos for verb, noun, adverb, preposition, regectedNonEnglishWord (#51)

* Trib 81: migrates all phrase related functionality from attribute service to phrase service (#52)

* TRIB-81: autowires phrase service and updates call on api/attributes to use phrase service

* TRIB-81: removes all phrase-related methods and word repositories

* TRIB-81: adds all phrase-related methods previously in attributes service. Also adds rejectedNonEnglishWord repo

---------
