+++ src/main/java/com/savvato/tribeapp/services/PhraseService.java
@@ -7,5 +7,9 @@ import java.util.Optional;
 
 public interface PhraseService {
 
+    boolean isPhraseValid(String verb, String noun, String adverb, String preposition);
+
+    void applyPhraseToUser(String verb, String noun, String adverb, String preposition);
+
     Optional<List<PhraseDTO>> getListOfPhraseDTOByUserId(Long id);
 }
